refactor: replace Node file reads in nav helper with `getCollection`
