+++ src/components/LeftSidebar/LeftSidebar.astro
@@ -1,7 +1,8 @@
 ---
 import Sponsors from './Sponsors.astro';
 import { removeLeadingSlash, removeTrailingSlash } from '../../util';
-import { getNav, useTranslations } from '../../i18n/util';
+import { getNav } from '../../util/getNav';
+import { useTranslations } from '../../i18n/util';
 import SidebarContent from './SidebarContent.astro';
 import SidebarToggleTabGroup from '../TabGroup/SidebarToggleTabGroup';
 import CommunityMenu from '../RightSidebar/CommunityMenu.astro';
